# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Cuisine Rest API with Flask",
        "version": "1.0.0",
        "title": "Cuisine Rest API with Flask",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "tags": [
        {
            "name": "Rest API for Cuisine",
            "description": "Cuisine Rest API"
        }
    ],
    "servers": [
        {
            "url": "/",
            "description": "HTTP"
        }
    ],
    "components": {
        "schemas": {
            "Cuisine": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "complete": {
                        "type": "boolean"
                    }
                }
            },
            "register": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "login": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "refresh": {
                "type": "object",
                "properties": {}
            },
            "logout": {
                "type": "object",
                "properties": {}
            },
            "Id": {
                "type": "integer",
                "properties": {
                    "id": {
                        "type": "integer"
                    }
                }
            },
            "cuisine_update": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "complete": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/cuisines": {
            "post": {
                "tags": [
                    "cuisines"
                ],
                "summary": "Add new cuisine",
                "requestBody": {
                    "description": "Send post request with data to add new cuisine",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cuisine"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "get": {
                "tags": [
                    "cuisines"
                ],
                "summary": "Get All cuisines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cuisines/{id}": {
            "get": {
                "tags": [
                    "cuisines"
                ],
                "summary": "Get Particular cuisine",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "provide cuisine id",
                        "required": true,
                        "type": "integer",
                        "schema": {
                            "$ref": "#/components/schemas/Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "cuisines"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "provide cuisine id",
                        "required": true,
                        "type": "integer",
                        "schema": {
                            "$ref": "#/components/schemas/Id"
                        }
                    }
                ],
                "summary": "Update cuisine",
                "requestBody": {
                    "description": "Update exixting cuisine",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/cuisine_update"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "cuisines"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "provide cuisine id",
                        "required": true,
                        "type": "integer",
                        "schema": {
                            "$ref": "#/components/schemas/Id"
                        }
                    }
                ],
                "summary": "Delete todo",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/components/schemas/Cuisine"
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "auths"
                ],
                "summary": "Add new User",
                "requestBody": {
                    "description": "Send post request with data to add new user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/register"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/components/schemas/register"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/components/schemas/register"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auths"
                ],
                "summary": "SignIn User",
                "requestBody": {
                    "description": "Send post request to signin user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/login"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Logged In",
                        "schema": {
                            "$ref": "#/components/schemas/login"
                        }
                    },
                    "404": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/components/schemas/login"
                        }
                    }
                }
            }
        },
        "/token/refresh": {
            "post": {
                "tags": [
                    "auths"
                ],
                "summary": "Refresh Token",
                "requestBody": {
                    "description": "Send post request to refresh token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/refresh"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged In",
                        "schema": {
                            "$ref": "#/components/schemas/refresh"
                        }
                    },
                    "404": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/components/schemas/refresh"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "auths"
                ],
                "summary": "Logout user",
                "requestBody": {
                    "description": "Send post request to logout user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/logout"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged out",
                        "schema": {
                            "$ref": "#/components/schemas/logout"
                        }
                    },
                    "404": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/components/schemas/logout"
                        }
                    }
                }
            }
        }
    }
}